auth-mdoule

+++ src/app/employee-list/employee-list.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+// import { dir } from 'console';
+import { NgModule } from '@angular/core';
 import { Employee } from '../employee';
 import { EmployeeService } from '../employee.service';
+
 
 @Component({
   selector: 'app-employee-list',
@@ -11,6 +14,9 @@ import { EmployeeService } from '../employee.service';
 export class EmployeeListComponent implements OnInit {
 
   employees: Employee[];
+  isDesc: boolean = false;
+  showData: any;
+
 
   // Injecting EmployeeService 
   constructor(private employeeService : EmployeeService,
@@ -32,14 +38,71 @@ export class EmployeeListComponent implements OnInit {
   }
 
   deleteEmployee(id : number){
-    this.employeeService.deleteEmployee(id).subscribe(data =>{
-      console.log(data);
-      this.getEmployees();
-    })
+
+    if(window.confirm('Are sure you want to delete this item ?')){
+      this.employeeService.deleteEmployee(id).subscribe(data =>{
+        console.log(data);
+        this.getEmployees();
+      })
+     }  
   }
 
   viewEmployee(id : number){
     this.router.navigate(['employee-details', id]);
   }
+
+  sortByName(property){
+    this.isDesc = !this.isDesc;
+    
+    let direction = this.isDesc ? 1 : -1;
+    this.employees.sort(function (a, b) {
+      if(a[property] < b[property]){
+        return -1 * direction;
+      }
+      else if(a[property] > b[property]){
+        return 1 * direction;
+      }
+      else{
+        return 0;
+      }
+    })
+  }
+
+  sortByEmail(property){
+    this.isDesc = !this.isDesc;
+    
+    let direction = this.isDesc ? 1 : -1;
+    this.employees.sort(function (a, b) {
+      if(a[property] < b[property]){
+        return -1 * direction;
+      }
+      else if(a[property] > b[property]){
+        return 1 * direction;
+      }
+      else{
+        return 0;
+      }
+    })
+  }
+
+  sortByLastName(lName){
+    this.isDesc = !this.isDesc;
+    
+    let direction = this.isDesc ? 1 : -1;
+    this.employees.sort(function (a, b) {
+      if(a[lName] < b[lName]){
+        return -1 * direction;
+      }
+      else if(a[lName] > b[lName]){
+        return 1 * direction;
+      }
+      else{
+        return 0;
+      }
+    })
+  }
+
+  
+
 
 }
